:sparkles: [#32] DELETE the bulkeditor (change to datagrid) and make it more generic
